make dbinder_transaction_data not depend on 32-bit/64-bit system

+++ ipc/native/c/rpc/src/dbinder_invoker.c
@@ -80,7 +80,7 @@ static void ToTransData(const IpcIo *data, dbinder_transaction_data *buf)
     buf->buffer_size = (data == NULL) ? 0 : (data->bufferCur - data->bufferBase);
     buf->offsets = buf->buffer_size;
     buf->offsets_size = (data == NULL) ? 0 :
-        ((size_t)(data->offsetsCur - data->offsetsBase) * sizeof(size_t));
+        (uint64_t)((data->offsetsCur - data->offsetsBase) * sizeof(size_t));
 }
 
 static void ToIpcData(const dbinder_transaction_data *tr, IpcIo *data)
@@ -159,7 +159,7 @@ static int32_t MoveTransData2Buffer(HandleSessionList *sessionObject, dbinder_tr
         RPC_LOG_ERROR("sessionObject buffer malloc failed");
         return ERR_FAILED;
     }
-    sessionObject->len = (uint32_t)transData->sizeOfSelf;
+    sessionObject->len = transData->sizeOfSelf;
 
     if (memcpy_s(sessionObject->buffer, sizeof(dbinder_transaction_data),
         transData, sizeof(dbinder_transaction_data)) != EOK) {
@@ -178,7 +178,7 @@ static int32_t MoveTransData2Buffer(HandleSessionList *sessionObject, dbinder_tr
     return ERR_NONE;
 }
 
-static HandleSessionList *WriteTransaction(int cmd, MessageOption option, int32_t handle,
+static HandleSessionList *WriteTransaction(int32_t cmd, MessageOption option, int32_t handle,
     int32_t sessionId, uint32_t code, IpcIo *data, uint64_t *seqNumber, int status)
 {
     HandleSessionList *sessionObject = GetSessionObject(handle, sessionId);
@@ -238,7 +238,7 @@ static int32_t OnSendMessage(HandleSessionList *sessionOfPeer)
     }
 
     int32_t ret = rpcSkeleton->rpcTrans->Send((int)sessionOfPeer->sessionId,
-        (void *)sessionOfPeer->buffer, (unsigned int)sessionOfPeer->len);
+        (void *)sessionOfPeer->buffer, (uint32_t)sessionOfPeer->len);
 
     free(sessionOfPeer->buffer);
     return ret;
@@ -594,7 +594,7 @@ void OnDatabusSessionClosed(int sessionId)
     }
 }
 
-static uint32_t HasCompletePackage(const char *data, uint32_t readCursor, unsigned int len)
+static uint32_t HasCompletePackage(const char *data, uint32_t readCursor, uint32_t len)
 {
     const dbinder_transaction_data *tr = (const dbinder_transaction_data *)(data + readCursor);
     if ((tr->magic == DBINDER_MAGICWORD) &&
@@ -605,7 +605,7 @@ static uint32_t HasCompletePackage(const char *data, uint32_t readCursor, unsign
     return 0;
 }
 
-void OnMessageAvailable(int sessionId, const void *data, unsigned int len)
+void OnMessageAvailable(int sessionId, const void *data, uint32_t len)
 {
     if (sessionId < 0 || data == NULL || len < sizeof(dbinder_transaction_data)) {
         RPC_LOG_ERROR("session has wrong inputs");
